# tmux-client/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: API for managing tmux sessions and windows
  title: tmux-client API
  version: 0.0.1-beta
servers:
- url: //localhost:8081/
paths:
...
  /v1/command/run:
    post:
      description: Executes a shell command
      operationId: runCommand
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tmux-client_internal_types.RunCommandRequest"
        description: Run command request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tmux-client_internal_types.RunCommandResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tmux-client_internal_types.APIError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tmux-client_internal_types.APIError"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tmux-client_internal_types.APIError"
          description: Internal Server Error
      summary: Run command
      tags:
      - Command
      x-codegen-request-body-name: request
...
components:
  schemas:
    tmux-client_internal_types.APIError:
      example:
        code: code
        details: details
        message: message
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
...
    tmux-client_internal_types.RunCommandRequest:
      properties:
        args:
          description: Arguments as separate items
          items:
            type: string
          type: array
        command:
          description: Executable name only
          type: string
        dry_run:
          description: "If true, don't actually run"
          type: boolean
        env:
          description: Additional env vars (KEY=VALUE)
          items:
            type: string
          type: array
        timeout:
          description: "Seconds, 0 = default (30s)"
          type: integer
        work_dir:
          description: Working directory
          type: string
      type: object
    tmux-client_internal_types.RunCommandResponse:
      example:
        args:
        - args
        - args
        duration_ms: 0
        stdout: stdout
        exit_code: 6
        dry_run: true
        timed_out: true
        stderr: stderr
        command: command
      properties:
        args:
          items:
            type: string
          type: array
        command:
          type: string
        dry_run:
          type: boolean
        duration_ms:
          type: integer
        exit_code:
          type: integer
        stderr:
          type: string
        stdout:
          type: string
        timed_out:
          type: boolean
      type: object
...